Move synchronous call implementations to base class
