+++ app/src/main/java/com/cryptoquack/model/exchange/BaseExchange.java
@@ -14,6 +14,7 @@ import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 
 import java.util.ArrayList;
+import java.util.Date;
 
 import io.reactivex.Single;
 
@@ -49,9 +50,12 @@ public abstract class BaseExchange {
 
     public abstract ArrayList<ExchangeMarket> getAvailableMarkets();
 
-    public abstract Double getCurrentPrice(ExchangeMarket market);
-
     public abstract Single<Double> getCurrentPriceAsync(ExchangeMarket market);
+
+    public Double getCurrentPrice(ExchangeMarket market) {
+        Single<Double> single = this.getCurrentPriceAsync(market);
+        return single.blockingGet();
+    }
 
     public abstract MonetaryAmount calculateFee(ExchangeAction.ExchangeActions action,
                                                 MonetaryAmount amount,
@@ -60,11 +64,39 @@ public abstract class BaseExchange {
     public abstract ArrayList<ExchangeAction.ExchangeActions> getAvailableActions(
             ExchangeMarket market);
 
-    public abstract Order makeOrder(Order orderRequest);
-
     public abstract Single<Order> makeOrderAsync(Order orderRequest);
 
-    public abstract OrderStatus getOrderStatus(Order order);
+    public Order makeOrder(Order orderRequest) {
+        Single<Order> single = this.makeOrderAsync(orderRequest);
+        return single.blockingGet();
+    }
 
     public abstract Single<OrderStatus> getOrderStatusAsync(Order order);
+
+    public OrderStatus getOrderStatus(Order order) {
+        Single<OrderStatus> single = this.getOrderStatusAsync(order);
+        return single.blockingGet();
+    }
+
+    public Single<ArrayList<Order>> getOrdersAsync(ExchangeMarket market, boolean liveOnly) {
+        return this.getOrdersAsync(market, null, null, liveOnly);
+    }
+
+    public ArrayList<Order> getOrders(ExchangeMarket market, boolean liveOnly) {
+        Single<ArrayList<Order>> single = this.getOrdersAsync(market, liveOnly);
+        return single.blockingGet();
+    }
+
+    public abstract Single<ArrayList<Order>> getOrdersAsync(ExchangeMarket market,
+                                                   final Date startTime,
+                                                   final Date endTime,
+                                                   boolean liveOnly);
+
+    public ArrayList<Order> getOrders(ExchangeMarket market,
+                                      final Date startTime,
+                                      final Date endTime,
+                                      boolean liveOnly) {
+        Single<ArrayList<Order>> single = this.getOrdersAsync(market, startTime, endTime, liveOnly);
+        return single.blockingGet();
+    }
 }
